[#22] automatically deactivate simple program change patches when another simple program change patch on the same MIDI channel is activated (disregards port)
Created a new patch type SimpleProgramChangePatch that can be explicitly defined by user, but is also automatically used when a normal patch that has only a single ProgramChange command is defined.  SimpleProgramChangePatches only support group A and only support a single ProgramChange command.  Normal patches that don't fit the requirements are not automatically converted to SimpleProgramChangePatch but some accommodation is made if a program change command is found in a normal patch definition.  SimpleProgramChangePatches do not cause active normal patches to become deactivated.   Only one SimpleProgramChangePatch is active per MIDI channel -- therefore up to 16 SimpleProgramChangePatches can be simultaneously active.

Use of SimpleProgramChangePatch means that exclusive switch groups in patchbanks do not need to be manually defined in the user config for program change commands targeting the same device.

This reduces the unintentional deactivation of normal type patches targeting one device when changing programs on other devices.